Simplify BuildingProperties into separate clasees
