+++ cea/demand/building_typology.py
@@ -1,0 +1,42 @@
+"""
+Building typology properties
+"""
+from __future__ import annotations
+from geopandas import GeoDataFrame as Gdf
+
+from cea.datamanagement.databases_verification import COLUMNS_ZONE_TYPOLOGY
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from cea.inputlocator import InputLocator
+
+
+class BuildingTypology:
+    """
+    Groups building typology properties used for the calc-thermal-loads functions.
+    """
+
+    def __init__(self, locator: InputLocator, building_names: list[str]):
+        """
+        Read building typology properties from input shape files and construct a new BuildingTypology object.
+
+        :param locator: an InputLocator for locating the input files
+        :param building_names: list of buildings to read properties for
+        """
+        if building_names is None:
+            building_names = locator.get_zone_building_names()
+
+        zone_gdf = Gdf.from_file(locator.get_zone_geometry())
+        self._prop_typology = zone_gdf[COLUMNS_ZONE_TYPOLOGY].set_index('name')
+        # Drop 'REFERENCE' column if it exists
+        if 'reference' in self._prop_typology:
+            self._prop_typology.drop('reference', axis=1, inplace=True)
+
+    def __getitem__(self, building_name: str) -> dict:
+        """Get typology properties of a building by name"""
+        return self._prop_typology.loc[building_name].to_dict()
+
+    def list_uses(self):
+        """get list of all uses (typology types)"""
+        return list(set(self._prop_typology['USE'].values))
